perf: refactor local file service
drop dependency on React
